pull types to their own file
